Move Load functions to controller
